requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
xgboost==1.7.6
scikit-learn==1.3.0
dnspython==2.4.2
requests==2.31.0
numpy==1.24.3